Halleluiah
dartboard is love, dartboard is life.

+++ projectile.js
@@ -5,6 +5,8 @@ class dartboard{
     this.currentY = -1;
     this.counter = 1;
     this.velocity = "y+";
+    this.speed = 0;
+    this.health = 0;
   }
 }
 
@@ -54,19 +56,27 @@ class Tile {
       this.element.appendChild(tower.element); // Append the tower element to the tile's DOM
       moneycount -= 15
       moneydisplay.innerHTML = `Money: $${moneycount}`;
-      const bullet = new Bullet(tower, getTarget());
+      const bullet = new Bullet(tower, getTarget(), towerType.seedPng, towerType.speed);
       bullet.hone();    
       function animate() {
+        console.log(".")
         bullet.target = getTarget();
         bullet.shoot();  // Update bullet position
         for (let enemy of enemies) {
           if (bullet.checkCollision(enemy)) {
-            enemy.health -=1;
             console.log(enemy.health)
-            if (enemy.health <= 0){
-              enemy.delete();
+            if (bullet.png == 'assets/RedBullet.png'){
+              enemy.health -=1;
+              if(enemy.health <= 0){
+                enemy.delete();
+                moneycount += 5;
+              }
             }
-            moneycount += 5;
+            if(bullet.png == 'assets/BlueBullet.png'){
+              enemy.speed = Math.max(enemy.speed*.75, 3);
+              bullet.target = getTarget();
+            }
+            
             moneydisplay.innerHTML = `Money: $${moneycount}`;
             
 
@@ -84,9 +94,11 @@ class Tile {
 }
 
 class TowerType{
-  constructor(type, png){
+  constructor(type, png, speed, seedPng){
     this.type = type;
     this.png = png;
+    this.speed = speed;
+    this.seedPng = seedPng;
   }
 
   displayInfo() {
@@ -99,7 +111,7 @@ var selectedTower = 1
 
 var tower1 = document.getElementById('tower1')
 var tower2 = document.getElementById('tower2')
-let towerType = new TowerType(1, 'assets/block.png');
+let towerType = new TowerType(1, 'assets/block.png', 20, 'assets/BlueBullet.png');
 
 tower1.addEventListener("click", function() {
   // Ensure both towers are reset
@@ -111,7 +123,7 @@ tower1.addEventListener("click", function() {
   tower2.classList.add("answerBtnsOff");
   
   selectedTower = 1;
-  towerType = new TowerType(selectedTower, 'assets/block.png');
+  towerType = new TowerType(selectedTower, 'assets/block.png', 20, 'assets/BlueBullet.png');
   towerType.displayInfo();
 });
 
@@ -125,7 +137,7 @@ tower2.addEventListener("click", function() {
   tower1.classList.add("answerBtnsOff");
   
   selectedTower = 2;
-  towerType = new TowerType(selectedTower, 'assets/block2.png');
+  towerType = new TowerType(selectedTower, 'assets/block2.png', 10, 'assets/RedBullet.png');
   towerType.displayInfo();
 });
 
@@ -148,9 +160,9 @@ tower2.addEventListener("mouseout", function() {
 });
 
 class Tower {
-  constructor(png, tile) {
+  constructor(type, png, tile) {
       
-      // this.type = type;
+      this.type = type;
       this.tile = tile; // The Tile object
       this.png = png; // The image for the tower
       this.element = this.createTowerElement(); // Create the DOM element for the tower
@@ -173,8 +185,9 @@ class Tower {
 
 // Bullet class definition
 class Bullet {
-    constructor(tower, enemy) {
+    constructor(tower, enemy, png, speed) {
       this.target = enemy;
+      this.png = png;
       this.spx = 0;
       this.spy = 0;
       this.homeX = tower.x;        // Initial X position (where the bullet starts)
@@ -182,13 +195,13 @@ class Bullet {
       this.currentX = tower.x;     // Current X position
       this.currentY = tower.y;     // Current Y position
       this.element = this.createBulletElement();  // Create the DOM element
-      this.overallSpeed = 20
+      this.overallSpeed = speed;
     }
   
     createBulletElement() {
       const bullet = document.createElement('div');
       bullet.classList.add('bullet');
-      bullet.style.backgroundImage = 'url(assets/crumb.png)';
+      bullet.style.backgroundImage = `url(${this.png})`;
       bullet.style.backgroundSize = 'cover';
       bullet.style.position = 'absolute';
       bullet.style.left = `${this.currentX}px`;
@@ -202,16 +215,16 @@ class Bullet {
       let y = this.target.currentY - this.homeY;
       
       const directionVelocity = this.target.velocity;
-      
+      let factorr = this.target.speed*10
       // Adjust x or y based on the direction
       if (directionVelocity === "y+") {
-          y += 60;
+          y += factorr;
       } else if (directionVelocity === "x+") {
-          x += 60;
+          x += factorr;
       } else if (directionVelocity === "y-") {
-          y -= 60;
+          y -= factorr;
       } else if (directionVelocity === "x-") {
-          x -= 60;
+          x -= factorr;
       }
       
       // Calculate the Euclidean distance (magnitude) between the points
@@ -246,7 +259,7 @@ class Bullet {
         this.element.style.top = `${this.currentY}px`;
 
         // Reset bullet if it goes off screen (you can adjust the left position reset based on the game area size)
-        if (this.currentX >  1300|| this.currentY >  1200|| this.currentX <  0|| this.currentY < 0) {
+        if (this.currentX >  1500|| this.currentY >  750|| this.currentX <  0|| this.currentY < 0) {
             this.currentX = this.homeX;
             this.currentY = this.homeY;
             this.hone();
@@ -319,6 +332,9 @@ class Enemy {
       this.delete()
       lives -= 1
       livesdisplay.innerHTML = `Lives: ${lives}`;
+      if(lives == 0){
+        window.location.replace("gameOverLose.html");
+      }
     }else{
       this.counter ++;
     }
@@ -334,17 +350,17 @@ let enemies = [];
 // Function to spawn a new enemy and add it to the enemies array
 function spawnEnemy(n) {
   if(n==1){
-  const enemy = new Enemy("assets/crumb.png", 330, 65, 15, 1);  // Starting position for each enemy
+  const enemy = new Enemy("assets/crumb.png", 330, 65, 12, 1);  // Starting position for each enemy
   enemies.push(enemy);
   console.log(enemies);
   }
   if(n==2){
-    const enemy = new Enemy("assets/crumb.png", 330, 65, 15, 2);  // Starting position for each enemy
+    const enemy = new Enemy("assets/crumb.png", 330, 65, 12, 2);  // Starting position for each enemy
     enemies.push(enemy);
     console.log(enemies);
     }
   if(n==3){
-    const enemy = new Enemy("assets/crumb.png", 330, 65, 30, 1);  // Starting position for each enemy
+    const enemy = new Enemy("assets/crumb.png", 330, 65, 24, 1);  // Starting position for each enemy
     enemies.push(enemy);
     console.log(enemies);
       }
@@ -522,7 +538,7 @@ fetch('finalmap.json')
         tileDiv.addEventListener('click', () => {
             console.log(`Tile ${tileId} clicked!`);
             if (tile.canPlaceTower()) {
-                const tower = new Tower(towerType.png, tile); // Create a new Tower
+                const tower = new Tower(towerType.type, towerType.png, tile); // Create a new Tower
                 tile.placeTower(tower); // Place the tower on the clicked tile
             } else {
                 console.log('Cannot place tower here!');
